split bindings into separate files
