+++ python/bindings/include/spark_dsg/python/scene_graph_iterators.h
