+++ python/bindings/src/misc_types.cpp
@@ -1,0 +1,85 @@
+/* -----------------------------------------------------------------------------
+ * Copyright 2022 Massachusetts Institute of Technology.
+ * All Rights Reserved
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ *  1. Redistributions of source code must retain the above copyright notice,
+ *     this list of conditions and the following disclaimer.
+ *
+ *  2. Redistributions in binary form must reproduce the above copyright notice,
+ *     this list of conditions and the following disclaimer in the documentation
+ *     and/or other materials provided with the distribution.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+ * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
+ * FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+ * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+ * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+ * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+ * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ *
+ * Research was sponsored by the United States Air Force Research Laboratory and
+ * the United States Air Force Artificial Intelligence Accelerator and was
+ * accomplished under Cooperative Agreement Number FA8750-19-2-1000. The views
+ * and conclusions contained in this document are those of the authors and should
+ * not be interpreted as representing the official policies, either expressed or
+ * implied, of the United States Air Force or the U.S. Government. The U.S.
+ * Government is authorized to reproduce and distribute reprints for Government
+ * purposes notwithstanding any copyright notation herein.
+ * -------------------------------------------------------------------------- */
+#include "spark_dsg/python/misc_types.h"
+
+#include <pybind11/pybind11.h>
+#include <spark_dsg/node_symbol.h>
+#include <spark_dsg/scene_graph_types.h>
+
+namespace py = pybind11;
+using namespace py::literals;
+
+namespace spark_dsg::python::misc_types {
+
+void addBindings(pybind11::module_& module) {
+  py::class_<DsgLayers>(module, "DsgLayers")
+      .def_readonly_static("SEGMENTS", &DsgLayers::SEGMENTS)
+      .def_readonly_static("OBJECTS", &DsgLayers::OBJECTS)
+      .def_readonly_static("AGENTS", &DsgLayers::AGENTS)
+      .def_readonly_static("PLACES", &DsgLayers::PLACES)
+      .def_readonly_static("MESH_PLACES", &DsgLayers::MESH_PLACES)
+      .def_readonly_static("STRUCTURE", &DsgLayers::STRUCTURE)
+      .def_readonly_static("ROOMS", &DsgLayers::ROOMS)
+      .def_readonly_static("BUILDINGS", &DsgLayers::BUILDINGS);
+
+  py::class_<LayerPrefix>(module, "LayerPrefix")
+      .def(py::init([](char key) { return LayerPrefix(key); }))
+      .def(py::init([](char key, uint32_t index) { return LayerPrefix(key, index); }))
+      .def(py::init([](uint32_t index) { return LayerPrefix(index); }))
+      .def_property(
+          "value",
+          [](const LayerPrefix& prefix) { return static_cast<uint32_t>(prefix); },
+          nullptr)
+      .def("__repr__", [](const LayerPrefix& prefix) { return prefix.str(true); });
+
+  py::class_<NodeSymbol>(module, "NodeSymbol")
+      .def(py::init([](char key, size_t index) { return NodeSymbol(key, index); }))
+      .def(py::init([](size_t value) { return NodeSymbol(value); }))
+      .def_property("category_id", &NodeSymbol::categoryId, nullptr)
+      .def_property("category", &NodeSymbol::category, nullptr)
+      .def_property(
+          "value",
+          [](const NodeSymbol& symbol) { return static_cast<NodeId>(symbol); },
+          nullptr)
+      .def("__repr__", &NodeSymbol::getLabel)
+      .def("__hash__",
+           [](const NodeSymbol& symbol) { return static_cast<NodeId>(symbol); })
+      .def(pybind11::self == pybind11::self)
+      .def(pybind11::self != pybind11::self);
+
+  py::implicitly_convertible<char, LayerPrefix>();
+}
+
+}  // namespace spark_dsg::python::misc_types
